requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
boto3==1.26.10
botocore==1.29.10
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==6.5.0
cssbeautifier==1.14.7
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-on-heroku==1.1.2
django-storages==1.13.1
djangorestframework==3.14.0
djlint==1.19.7
drf-spectacular==0.24.2
EditorConfig==0.12.3
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
jmespath==1.0.1
jsbeautifier==1.14.7
jsonschema==4.16.0
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.2
psycopg2-binary==2.9.5
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-mock==1.10.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.10.0
